backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework-simplejwt==5.2.2

# Database and Channels
channels-redis==4.1.0

# Environment and Security
python-dotenv==1.0.0
daphne==4.0.0
python-magic

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.1.12
langchain-core==0.1.39
pydantic>=2.0.0
openai==1.3.0
numpy==1.24.3

# Google Gemini (Generative AI)
google-generativeai==0.3.2
langchain-google-genai==0.0.7

# Development Tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0

# Authentication
google-auth==2.30.0

# Payments
stripe==8.7.0

# Production Deployment
dj-database-url==1.3.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.6.0 